Add files via upload
It contains 8X3 priority encoder that takes 8-bit input (in) and generates 3-bit output (out). When multiple bits are high, it assigns the output based on the highest-priority, starting from in[0] (lowest index) to in[7] (highest index). The encoder outputs the position of the first high bit, with output set to the index of that bit. If none of the bits are set, the output will not change as there is no default assignment for cases when it is all zeroes.